wapibot/requirements.txt
# Core Framework Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-extra-types >= 2.10.6
phonenumbers>=9.0.21
email-validator>=2.3.0

# DSPy and LLM Dependencies
dspy-ai>=3.0.0
litellm>=1.40.0

# HTTP Client and Async Support
httpx>=0.25.0
aiofiles>=23.2.0

# Data Processing and Validation
python-dateutil>=2.8.2
pytz>=2023.3

# WhatsApp Integration (Optional)
# pywa>=1.0.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment and Configuration
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.2.0

# Development Dependencies (Optional)
ruff>=0.1.0
black>=23.0.0
mypy>=1.7.0
pyrefly >= 0.46.0

# Additional Utilities
requests>=2.31.0

# WapiBot Backend V2 - Additional Dependencies
# Only packages NOT in root requirements.txt

# ============================================================================
# FASTAPI EXTRAS
# ============================================================================
python-multipart>=0.0.12
pydantic-settings>=2.6.0

# ============================================================================
# LANGGRAPH WORKFLOW ORCHESTRATION
# ============================================================================
langgraph>=0.2.40
langgraph-checkpoint-sqlite>=2.0.0
langchain-core>=0.3.0

# ============================================================================
# LLM PROVIDERS
# ============================================================================
ollama>=0.4.0
openai>=1.50.0  # Optional: for OpenAI models

# ============================================================================
# DATABASE & PERSISTENCE
# ============================================================================
sqlmodel>=0.0.27
aiosqlite>=0.20.0

# ============================================================================
# TESTING EXTRAS
# ============================================================================
pytest-cov>=6.0.0

# ============================================================================
# CODE QUALITY EXTRAS
# ============================================================================
isort>=5.13.0
pylint>=3.3.0


# UPI
qrcode[pil]>=7.4.2

# ============================================================================
# CELERY & TASK QUEUE
# ============================================================================
celery>=5.4.0
redis>=5.0.0
#=============================================================================
# Security
#=============================================================================
cryptography>=46.0.3
jwt>=1.4.0

# Tunnelling

pyngrok>=7.5.0